feat(DataSpreadsheet): update selection area and active cell interaction
