+++ packages/cloud-cognitive/src/components/DataSpreadsheet/DataSpreadsheet.js
@@ -14,7 +14,6 @@ import React, {
   useEffect,
 } from 'react';
 import { useBlockLayout, useTable } from 'react-table';
-import { px } from '@carbon/layout';
 
 // Other standard imports.
 import PropTypes from 'prop-types';
@@ -29,12 +28,14 @@ import { getDevtoolsProps } from '../../global/js/utils/devtools';
 import { getScrollbarWidth } from '../../global/js/utils/getScrollbarWidth';
 import { useActiveElement, usePreviousValue } from '../../global/js/hooks';
 import uuidv4 from '../../global/js/utils/uuidv4';
+import { deepCloneObject } from '../../global/js/utils/deepCloneObject';
 
 import {
   useResetSpreadsheetFocus,
   useSpreadsheetOutsideClick,
   useMoveActiveCell,
   useMultipleKeyTracking,
+  useSpreadsheetEdit,
 } from './hooks';
 
 import { createActiveCellFn } from './utils/createActiveCellFn';
@@ -47,6 +48,7 @@ import {
 import { handleHeaderCellSelection } from './utils/handleHeaderCellSelection';
 import { removeCellSelections } from './utils/removeCellSelections';
 import { selectAllCells } from './utils/selectAllCells';
+import { handleEditSubmit } from './utils/handleEditSubmit';
 // cspell:words rowcount colcount
 
 // The block part of our conventional BEM class names (blockClass__E--M).
@@ -663,132 +665,67 @@ export let DataSpreadsheet = React.forwardRef(
       startEditMode();
     };
 
-    const updateSelectionAreaOnCellEditSubmit = ({ type }) => {
-      const submitEditChanges = () => {
-        const prevCoords = previousState?.activeCellCoordinates;
-        const cellProps = rows[prevCoords?.row].cells[prevCoords?.column];
-        removeCellEditor();
-        updateData(prevCoords?.row, cellProps.column.id);
-      };
-      removeCellSelections({ spreadsheetRef });
-      submitEditChanges();
-      const tempMatcher = uuidv4();
-      const newSelectionArea = {
-        row:
-          type === 'Enter'
-            ? activeCellCoordinates.row === rows.length - 1
-              ? activeCellCoordinates.row
-              : activeCellCoordinates.row + 1
-            : activeCellCoordinates.row,
-        column:
-          type === 'Tab'
-            ? activeCellCoordinates.column === columns.length - 1
-              ? activeCellCoordinates.column
-              : activeCellCoordinates.column + 1
-            : activeCellCoordinates.column,
-      };
-      setSelectionAreas([
-        {
-          point1: newSelectionArea,
-          point2: newSelectionArea,
-          matcher: tempMatcher,
-          areaCreated: false,
-        },
-      ]);
-      setCurrentMatcher(tempMatcher);
-      cellEditorRulerRef.current.textContent = '';
-    };
-
-    // Update the data
-    const handleEditSubmit = (event) => {
-      const { key } = event;
-      if (key === 'Enter') {
-        updateSelectionAreaOnCellEditSubmit({ type: 'Enter' });
-        setActiveCellCoordinates((prev) => ({
-          ...prev,
-          row: prev.row === rows.length - 1 ? prev.row : prev.row + 1, // do not move to next cell below if we're already in the last row
-        }));
-      }
-      if (key === 'Tab') {
-        event.preventDefault();
-        updateSelectionAreaOnCellEditSubmit({ type: 'Tab' });
-        setActiveCellCoordinates((prev) => ({
-          ...prev,
-          column:
-            prev.column === columns.length - 1 ? prev.column : prev.column + 1, // do not move to next cell below if we're already in the last column
-        }));
-      }
-      return;
-    };
-
-    useEffect(() => {
-      if (isEditing) {
-        const cellProps =
-          rows[activeCellCoordinates?.row]?.cells[
-            activeCellCoordinates?.column
-          ];
-        const activeCellLeftPosition = activeCellRef?.current.style.left;
-        const activeCellTopPosition = activeCellRef?.current.style.top;
-        cellEditorRef.current.style.left = activeCellLeftPosition;
-        cellEditorRef.current.style.top = activeCellTopPosition;
-        cellEditorRef.current.style.display = 'block';
-        cellEditorRef.current.style.width = activeCellRef?.current.style.width;
-        cellEditorRef.current.style.height =
-          activeCellRef?.current.style.height;
-        cellEditorRef.current.style.paddingTop = `${
-          (parseInt(activeCellRef?.current.style.height) - 16) / 2 - 1
-        }px`; // calculate paddingTop based on cellHeight which could be variable depending on the cellSize prop
-        cellEditorRef.current.style.textAlign =
-          cellProps?.column?.placement === 'right' ? 'right' : 'left';
-        cellEditorRef.current?.focus();
-        const rulerWidth = cellEditorRulerRef.current.offsetWidth;
-        const cellWidth = activeCellRef.current.offsetWidth;
-        if (rulerWidth >= cellWidth) {
-          const widthMultiplier = Math.floor(rulerWidth / cellWidth) + 1;
-          const startingColumnPosition = activeCellCoordinates?.column;
-          const startingRowPosition = activeCellCoordinates?.row;
-          const totalColumns = columns.length;
-          const totalRows = rows.length;
-          const totalMultiplierPossible = totalColumns - startingColumnPosition;
-          const totalCellEditorMaxHeight =
-            (totalRows - startingRowPosition) * defaultColumn.rowHeight;
-          cellEditorRef.current.style.maxHeight = px(totalCellEditorMaxHeight);
-          cellEditorRef.current.style.width = px(
-            cellWidth *
-              (widthMultiplier <= totalMultiplierPossible
-                ? widthMultiplier
-                : totalMultiplierPossible)
-          );
-          cellEditorRef.current.style.height = px(
-            cellEditorRef.current.scrollHeight
-          ); // adds dynamic height to cell editor
-          // Cell editor has reached max height, we need to add the scrolling back.
-          // We also need to subtract 1 to account for the fact that the cell editor
-          // is placed one pixel below the cell being edited to account for the border
-          if (
-            cellEditorRef.current.clientHeight ===
-            totalCellEditorMaxHeight - 1
-          ) {
-            cellEditorRef.current.style.overflow = 'auto';
-          } else {
-            cellEditorRef.current.style.overflow = 'hidden';
-          }
-        }
-      }
-      if (!isEditing) {
-        cellEditorRef.current.style.overflow = 'hidden';
-        cellEditorRef.current.style.display = 'none';
-        cellEditorRef.current.blur();
-        activeCellRef.current.focus();
-      }
-    }, [
+    useSpreadsheetEdit({
       isEditing,
-      activeCellCoordinates,
       rows,
-      cellEditorValue,
-      columns.length,
+      activeCellCoordinates,
+      activeCellRef,
+      cellEditorRef,
+      cellEditorRulerRef,
+      columns,
       defaultColumn,
+      cellEditorValue,
+    });
+
+    const handleActiveCellMouseEnter = useCallback(() => {
+      handleActiveCellMouseEnterCallback(selectionAreas, clickAndHoldActive);
+    }, [
+      clickAndHoldActive,
+      selectionAreas,
+      handleActiveCellMouseEnterCallback,
     ]);
+
+    // Only update if there are cell selection areas
+    // Find point object that matches currentMatcher and remove the second point
+    // because hovering over the active cell while clicking and holding should
+    // remove the previously existing selection area
+    const handleActiveCellMouseEnterCallback = useCallback(
+      (areas, clickHold) => {
+        if (!currentMatcher) {
+          return;
+        }
+        if (areas && areas.length && clickHold && currentMatcher) {
+          setSelectionAreas((prev) => {
+            const selectionAreaClone = deepCloneObject(prev);
+            const indexOfItemToUpdate = selectionAreaClone.findIndex(
+              (item) => item.matcher === currentMatcher
+            );
+            if (indexOfItemToUpdate === -1) {
+              return prev;
+            }
+            if (
+              typeof selectionAreaClone[indexOfItemToUpdate].point2 ===
+                'object' &&
+              selectionAreaClone[indexOfItemToUpdate].areaCreated
+            ) {
+              selectionAreaClone[indexOfItemToUpdate].point2 =
+                selectionAreaClone[indexOfItemToUpdate].point1;
+              selectionAreaClone[indexOfItemToUpdate].areaCreated = false;
+              activeCellRef.current.classList.remove(
+                `${blockClass}__active-cell--with-selection`
+              );
+              removeCellSelections({
+                matcher: currentMatcher,
+                spreadsheetRef,
+              });
+              return selectionAreaClone;
+            }
+            return prev;
+          });
+        }
+      },
+      [spreadsheetRef, currentMatcher]
+    );
 
     return (
       <div
@@ -863,6 +800,7 @@ export let DataSpreadsheet = React.forwardRef(
             onClick={handleActiveCellClick}
             onKeyDown={handleActiveCellKeyDown}
             onDoubleClick={handleActiveCellDoubleClick}
+            onMouseEnter={handleActiveCellMouseEnter}
             ref={activeCellRef}
             className={cx(
               `${blockClass}--interactive-cell-element`,
@@ -874,7 +812,19 @@ export let DataSpreadsheet = React.forwardRef(
           </button>
           <TextArea
             value={cellEditorValue}
-            onKeyDown={handleEditSubmit}
+            onKeyDown={handleEditSubmit({
+              activeCellCoordinates,
+              cellEditorRulerRef,
+              columns,
+              previousState,
+              removeCellEditor,
+              rows,
+              setActiveCellCoordinates,
+              setCurrentMatcher,
+              setSelectionAreas,
+              spreadsheetRef,
+              updateData,
+            })}
             onChange={(event) => {
               setCellEditorValue(event.target.value);
               cellEditorRulerRef.current.textContent = event.target.value;
