+++ packages/cloud-cognitive/src/components/DataSpreadsheet/_data-spreadsheet.scss
@@ -168,6 +168,19 @@
         outline: 0;
       }
     }
+    .#{$block-class}__active-cell--highlight.#{$block-class}__active-cell--with-selection {
+      &::before {
+        position: absolute;
+        top: 0;
+        left: 0;
+        display: block;
+        width: 100%;
+        height: 100%;
+        background-color: $interactive-01;
+        content: '';
+        opacity: 0.25;
+      }
+    }
     .#{$block-class}__selection-area--element {
       position: absolute;
       z-index: 2;
